refactor : interface 내의 메소드 제거
