+++ src/main/java/com/seungwon/springbooturlshortener/application/UrlShortenService.java
@@ -3,17 +3,5 @@ package com.seungwon.springbooturlshortener.application;
 import com.seungwon.springbooturlshortener.domain.Url;
 
 public interface UrlShortenService {
-
-	public static UrlShortenService from(String strategy) {
-		switch (strategy) {
-			case "random" -> {
-				return new RandomUrlShortenService();
-			}
-			default -> {
-				return new Base64UrlShortenService();
-			}
-		}
-	}
-
 	String shorten(Url original);
 }
